src/pyproject.toml
[project]
name = "concurp2-lab22"
version = "0.1.0"
description = "An asyncio-based web image scraper."
authors = [
    {name = "Noam Maissen", email = "noam.maissen@edu.hefr.ch"},
    {name = "Baptiste Roland", email = "baptiste.roland@edu.hefr.ch"},
]
dependencies = [
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.5",
    "beautifulsoup4>=4.12.3",
    "pillow>=10.3.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.poetry]
name = "image_scraper"
version = "0.1.0"
description = "An asyncio-based web image scraper."
authors = ["Noam Maissen <noam.maissen@edu.hefr.ch>", "Baptiste Roland <baptiste.roland@edu.hefr.ch"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "^23.2.1"
beautifulsoup4 = "^4.12.3"
pillow = "^10.3.0"
aiohttp = "^3.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"